# internal/ports/chi/router.go
package chi

import (
	"encoding/json"
	dto "final/pkg"
	"fmt"
	"log"
	"net/http"
	"strings"
	"time"

	httpSwagger "github.com/swaggo/http-swagger"

	"github.com/go-chi/chi/v5"
)

type Server struct {
	router  *chi.Mux
	service PriceUseCase
}

func (s *Server) ServeHTTP(w http.ResponseWriter, req *http.Request) {
	s.router.ServeHTTP(w, req)
}
func NewServer(service PriceUseCase) (*Server, error) {
	if service == nil { //
		return nil, fmt.Errorf("errprs") // хиитрую ошибку добавить
	}
	r := chi.NewRouter()

	return &Server{router: r,
		service: service,
	}, nil
}

func (s *Server) Start() {
	s.router.Get("/get/prices/last", s.GetLastPrice)
	s.router.Get("/get/prices/min", s.GetMinPrice)
	s.router.Get("/get/prices/max", s.GetMaxPrice)
	s.router.Get("/get/prices/percent", s.GetChangePercent)

	// Swagger UI
	s.router.Get("/swagger/*", httpSwagger.WrapHandler)

	// Swagger JSON
	s.router.Get("/swagger/doc.json", func(w http.ResponseWriter, r *http.Request) {
		http.ServeFile(w, r, "./docs/swagger.json")
	})

	log.Println("Сервер запущен на http://localhost:8080")
	log.Fatal(http.ListenAndServe(":8080", s))
}

// ================GetLastPrice================//
// @Summary Последние цены
// @Param symbols query string true "Символы через запятую"
// @Success 200 {array} dto.PriceDTO
// @Failure 400 {string} string "Bad request"
// @Failure 404 {string} string "Not found"
// @Failure 500 {string} string "Internal error"
// @Router /get/prices/last [get]
func (s *Server) GetLastPrice(rw http.ResponseWriter, req *http.Request) {

	symbols := req.URL.Query().Get("symbols")
	if symbols == "" {
		http.Error(rw, "symbols is required", http.StatusBadRequest)
		return
	}
	splitSymbols := strings.Split(symbols, ",")
	prices, err := s.service.GetPricesLast(req.Context(), splitSymbols)
	if err != nil {
		log.Printf("ERROR: %v", err)
		http.Error(rw, "internal error", http.StatusInternalServerError)
		return
	}
	if len(prices) == 0 {
		http.Error(rw, "not found", http.StatusNotFound)
		return
	}

	var data []dto.PriceDTO
	for _, v := range prices {
		data = append(data, dto.PriceDTO{
			Symbol: v.Symbol,
			Price:  v.Price,
			Time:   v.CreatedAt.Format(time.RFC3339),
		})
	}

	json.NewEncoder(rw).Encode(data)
}

// ================GetMaxPrice================//
// @Summary Максимальные цены
// @Param symbols query string true "Символы через запятую"
// @Success 200 {array} dto.PriceDTO
// @Failure 400 {string} string "Bad request"
// @Failure 404 {string} string "Not found"
// @Failure 500 {string} string "Internal error"
// @Router /get/prices/max [get]
func (s *Server) GetMaxPrice(rw http.ResponseWriter, req *http.Request) {

	symbols := req.URL.Query().Get("symbols")
	if symbols == "" {
		http.Error(rw, "symbols is required", http.StatusBadRequest)
		return
	}
	splitSymbols := strings.Split(symbols, ",")
	prices, err := s.service.GetMaxPrices(req.Context(), splitSymbols)
	if err != nil {
		log.Printf("ERROR: %v", err)
		http.Error(rw, "internal error", http.StatusInternalServerError)
		return
	}
	if len(prices) == 0 {
		http.Error(rw, "not found", http.StatusNotFound)
		return
	}

	var data []dto.PriceDTO
	for _, v := range prices {
		data = append(data, dto.PriceDTO{
			Symbol: v.Symbol,
			Price:  v.Price,
			Time:   v.CreatedAt.Format(time.RFC3339),
		})
	}

	json.NewEncoder(rw).Encode(data)
}

// ================GetMinPrice================//
// @Summary Минимальные цены
// @Param symbols query string true "Символы через запятую"
// @Success 200 {array} dto.PriceDTO
// @Failure 400 {string} string "Bad request"
// @Failure 404 {string} string "Not found"
// @Failure 500 {string} string "Internal error"
// @Router /get/prices/min [get]
func (s *Server) GetMinPrice(rw http.ResponseWriter, req *http.Request) {

	symbols := req.URL.Query().Get("symbols")
	if symbols == "" {
		http.Error(rw, "symbols is required", http.StatusBadRequest)
		return
	}
	splitSymbols := strings.Split(symbols, ",")
	prices, err := s.service.GetMinPrices(req.Context(), splitSymbols)
	if err != nil {
		log.Printf("ERROR: %v", err)
		http.Error(rw, "internal error", http.StatusInternalServerError)
		return
	}
	if len(prices) == 0 {
		http.Error(rw, "not found", http.StatusNotFound)
		return
	}

	var data []dto.PriceDTO
	for _, v := range prices {
		data = append(data, dto.PriceDTO{
			Symbol: v.Symbol,
			Price:  v.Price,
			Time:   v.CreatedAt.Format(time.RFC3339),
		})
	}

	json.NewEncoder(rw).Encode(data)
}

// ================GetChangePrices================//
// @Summary Изменение в процентах
// @Param symbols query string true "Символы через запятую"
// @Success 200 {array} dto.PriceDTO
// @Failure 400 {string} string "Bad request"
// @Failure 404 {string} string "Not found"
// @Failure 500 {string} string "Internal error"
// @Router /get/prices/percent [get]
func (s *Server) GetChangePercent(rw http.ResponseWriter, req *http.Request) {

	symbols := req.URL.Query().Get("symbols")
	if symbols == "" {
		http.Error(rw, "symbols is required", http.StatusBadRequest)
		return
	}
	splitSymbols := strings.Split(symbols, ",")
	prices, err := s.service.GetChangePercent(req.Context(), splitSymbols)
	if err != nil {
		log.Printf("ERROR: %v", err)
		http.Error(rw, "internal error", http.StatusInternalServerError)
		return
	}
	if len(prices) == 0 {
		http.Error(rw, "not found", http.StatusNotFound)
		return
	}

	var data []dto.PriceDTO
	for _, v := range prices {
		data = append(data, dto.PriceDTO{
			Symbol: v.Symbol,
			Price:  v.Price,
			Time:   v.CreatedAt.Format(time.RFC3339),
		})
	}

	json.NewEncoder(rw).Encode(data)
}